requirements.txt
# Raspberry Pi Sensor Monitoring System Requirements

# GPIO Control
RPi.GPIO==0.7.1

# NFC/FeliCa Reader
nfcpy==1.0.4

# Temperature/Humidity Sensor (DHT22/AM2302)
Adafruit-DHT==1.4.0

# Alternative DHT library (if Adafruit-DHT doesn't work)
# Adafruit-CircuitPython-DHT==3.7.7

# GUI Framework (already included in Python)
# tkinter - built-in

# Logging and utilities (built-in)
# datetime, time, threading, queue, json, logging

# Optional: Data visualization
# matplotlib==3.5.3
# numpy==1.21.6

# Optional: Data storage
# sqlite3 - built-in
# pandas==1.3.5

# Optional: Web interface
# flask==2.0.3
# flask-socketio==5.1.1

# Optional: MQTT communication
# paho-mqtt==1.6.1

# Development dependencies
# pytest==7.1.3
# black==22.3.0
# flake8==4.0.1